==> source_code/Cafe/광진구/4-1. normalized_matrix.ipynb <==
import pandas as pd
import numpy as np
import math

# get size of vector
def mag(x):
    return math.sqrt(sum(i**2 for i in x))

res_matrix = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_res_matrix.csv')
user_matrix = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_user_matrix.csv')
group_count = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_group_count.csv',names=['group','count'])


res_matrix

group_count.drop([0], inplace =True)

res_matrix.rename(columns={'Unnamed: 0':'p_id'}, inplace=True)
res_matrix.set_index('p_id', drop=True, inplace=True)
group_count.set_index('group', drop=True, inplace=True)

user_matrix

user_matrix.rename(columns={'Unnamed: 0':'u_id'}, inplace=True)
# user_matrix.set_index('u_id', drop=True, inplace=True)


user_matrix

user_matrix.set_index('u_id', drop=True, inplace=True)

user_matrix

res_matrix


group_count.head()

p_m = pd.DataFrame(np.array(res_matrix) * (1/np.array(group_count['count'])), index=res_matrix.index)
p_m = p_m.iloc[:,:].div(p_m.sum(axis=1), axis=0).fillna(0)
u_m = pd.DataFrame(np.array(user_matrix) * (1/np.array(group_count['count'])), index=user_matrix.index)
u_m = u_m.iloc[:,:].div(u_m.sum(axis=1), axis=0).fillna(0)

p_m

u_m

p_m.to_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_normalized_item_matrix.csv')
u_m.to_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_normalized_user_matrix.csv')

group_count.to_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_group_count.csv')
