==> source_code/Cafe/광진구/6-1. CBF_rmse_광진구_cafe.ipynb <==
import pandas as pd
import numpy as np
import math
from tqdm import tqdm

# get size of vector
def mag(x):
    return math.sqrt(sum(i**2 for i in x))

p_m = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_normalized_item_matrix.csv')
u_m = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_normalized_user_matrix.csv')
group_count = pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_group_count.csv')

# res_matrix.rename(columns ={'Unnamed: 0':'p_id' }, inplace =True)
p_m.set_index('p_id', inplace =True)
# user_matrix.rename(columns ={'Unnamed: 0':'u_id' }, inplace =True)
u_m.set_index('u_id', inplace =True)
# group_count.rename(columns ={'Unnamed: 0':'group' }, inplace =True)
# group_count.set_index('group', inplace =True)
# res_matrix.set_index('p_id', drop=True, inplace=True)
# user_matrix.set_index('u_id', drop=True, inplace=True)
# group_count.set_index('group', drop=True, inplace=True)

from google.colab import drive
drive.mount('/content/drive')

group_count

group_count.set_index('group', inplace =True)

group_count

p_m

u_m

from sklearn.metrics.pairwise import cosine_similarity

def cos_sim_matrix(a,b):
  cos_sim=cosine_similarity(a,b)
  result_df =pd.DataFrame(data=cos_sim, index=[a.index])

  return result_df

#similarity df


cs_u_df = cos_sim_matrix(u_m,u_m)
cs_u_df.head()

cs_p_df = cos_sim_matrix(p_m,p_m)
cs_p_df.head()

eat_review =pd.read_csv('/content/drive/MyDrive/Colab Notebooks/data/광진구_cafe_review.csv')
eat_review

eat_review['u_id'].value_counts()

eat_review['rating'].value_counts()

len(eat_review['p_id'].unique())

eat_review.drop(['review','res_name','user_name'],axis=1, inplace=True)

eat_review

eat_review_mean=eat_review.groupby(['u_id','p_id']).mean()
eat_review_mean

eat_review_mean.reset_index(level=0, inplace=True)
eat_review_mean.reset_index(level=0, inplace=True)
eat_review_mean

eat_review_mean.shape

# eat_review_mean_dropDU=eat_review_mean.drop_duplicates(subset=['u_id','p_id'], keep='first', ignore_index=False)

# eat_review_mean_dropDU.shape

import tensorflow as tf

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error

train_df, test_df = train_test_split(eat_review_mean, test_size=0.2, random_state=123)

train_df.shape

test_userids =list(set(test_df['u_id'].values))
len(test_userids)

all_userids =list(set(eat_review_mean['u_id'].values))
len(all_userids)#############################

user_record_df=eat_review[eat_review['u_id']==1567]

user_record_df

user_record_df['p_id']

#user record df

user_sim_df= cs_p_df.loc[user_record_df['p_id']]
user_sim_df

 user_record_df=train_df.loc[train_df['u_id']==int(1567),:]
 user_record_df

user_sim_df= cs_u_df.loc[user_record_df['p_id']]
user_sim_df

user_rating_df = user_record_df[['rating']] 
user_rating_df

#prediction

 prediction =np.matmul(user_sim_df.T.to_numpy() , user_rating_df.to_numpy()).flatten()
 prediction

result_df =pd.DataFrame()

for user_id in tqdm(all_userids):
  user_record_df = eat_review_mean.loc[eat_review_mean['u_id']==int(user_id),:] # (n,3) 

  user_sim_df= cs_p_df.loc[user_record_df['p_id']] #(n,1730). #1730개......##A
  user_rating_df = user_record_df[['rating']] #(n,1)##B
  sim_sum = np.sum(user_sim_df.T.to_numpy(),-1)


  prediction =np.matmul(user_sim_df.T.to_numpy() , user_rating_df.to_numpy()).flatten() /(sim_sum+1)##AB
  
  prediction_df = pd.DataFrame(prediction , index = cs_p_df.index).reset_index()
  prediction_df.columns=['p_id', 'pred_rating']
  prediction_df = prediction_df[['p_id', 'pred_rating']][prediction_df['p_id'].isin(test_df[test_df['u_id'] == user_id]['p_id'].values)]

  temp_df = prediction_df.merge(test_df[test_df['u_id'] == user_id], on='p_id')
  result_df = pd.concat([result_df, temp_df], axis=0)

result_df.head(10)


result_df.shape

result_df['pred_rating'].mean()

result_df['rating'].mean()


mse = mean_squared_error(y_true=result_df['rating'].values, y_pred=result_df['pred_rating'].values)
rmse = np.sqrt(mse)

print(mse, rmse)
